Fixing the units of qpoints for Quantum ESPRESSO
In the matdyn.modes, the qpoints are stored in units of
2*pi/alat, where alat is a "lattice parameter" defined
in the Quantum ESPRESSO (QE) run.
This value might not be defined in the input SCF file,
in particular in the case `ibrav=0`. In this case, recent
versions of QE set `alat` equal to the length of the first
lattice vector, but earlier 5.x versions had a different
behavior.
If `ibrav` is not 0, instead, `alat` is set from either
`celldm` (in bohr) or `a` (in angstrom) from the input.

Therefore, there is no reliable way to infer `alat` from
the input. Therefore, the only solution is to ask also to
pass one of the output files (both the pw.x and the ph.x
files contain the information on the `alat`).
This is now parsed, converted from bohr to angstrom, and
used internally to correctly convert qpoints from units
of 2pi/alat in units of 1/angstrom, and then in reduced
coordinates as needed internally.

Note that this issue only exists when parsing Quantum ESPRESSO
files.
Now, for double checking, when the phonons come from a
Quantum ESPRESSO run, we also print the value of alat to the
output HTML page for double checking (NOTE! there is no
consistency check performed at this point that the input
and output files provided are from the same run - only a
warning is shown for the user to double check)

The bug had been already reported in #5 some months ago,
and then reported again by Chiara Cignarella (EPFL) a few
days ago - this fix has been tested together with her.

This fixes #5