Limiting framerate
The FPS were limited to 60FPS, but at every click
on a button or on the phonon band structure, the loop
creating the frames was called again, generating 2, 3, ...
more loops running in parallel, effectively limiting
the FPS only to 120, 180, ... - i.e., in practice,
limiting at the max allowed by the CPU in most cases.

Now I remove the calles to animate() in a few places like
the update button of the supercell size, but also I
wrap the animate function (calling itself with framerate
limiting) inside an internal `animate_internal` function,
that is protected by a boolean variable, so it's called
only once.

Fixes #1