Extracting the SCF via the phonons from MC
I update the script so it gets the SCF input via the parents of the phonons,
rather than from the parents of the bands, that sometimes are just vc-relax.

I also store a UUID.
The script now fetches all existing 2D phonons.
These are not stored here, only the examples are updated with the results
of this updated script.

Note: the script could very much be improved by implementing a general
class to traverse the graph via the command-line interface, but at the
moment this is too much work so I did a lot of copy-paste of the code.